refactor(galgame): split plugin_entries/ into 39 mixin files
Extract each @plugin_entry method from GalgamePlugin into its own mixin file
under plugin_entries/, then mount them on GalgamePlugin via multiple
inheritance. Method bodies are preserved verbatim (no renames, no logic
changes, no comment edits).

- 39 mixin files in plugin_entries/ (one per @plugin_entry method)
- plugin_entries/_common.py centralizes shared imports with explicit __all__
  to re-export underscore-prefixed private helpers across star-import
- plugin_constants.py moved 12 module-level constants out of __init__.py so
  both the class body and mixins import them from a neutral module
- GalgamePlugin now inherits from 39 mixins + NekoPluginBase
- 2 @lifecycle + 1 @timer_interval methods stay inline (to move to plugin_core.py in next commit)

Validated: inspect.getmembers_static walks the MRO and finds all 42 entries
(39 action + 2 lifecycle + 1 timer).
