+++ plugin/plugins/galgame_plugin/__init__.py
@@ -158,20 +158,20 @@ from .plugin_util_helpers import (
     _open_url_in_browser,
 )
 
-_OCR_BACKEND_SELECTIONS = {"auto", "rapidocr"}
-_OCR_CAPTURE_BACKEND_SELECTIONS = {"auto", "smart", "dxcam", "mss", "pyautogui", "printwindow"}
-_BACKGROUND_BRIDGE_POLL_MIN_STALE_SECONDS = 45.0
-_BRIDGE_TICK_INTERVAL_SECONDS = 1.0
-# Foreground refresh TTL: repeated calls within two seconds return early so
-# bridge_tick, advance monitor, and status payload refreshes stay idempotent.
-_OCR_FOREGROUND_REFRESH_TTL_SECONDS = 2.0
-_LATENCY_SAMPLE_LIMIT = 120
-_LATENCY_MIN_SAMPLES_FOR_P95 = 5
-_OCR_POLL_P95_DEGRADE_THRESHOLD_SECONDS = 3.0
-_OCR_FOREGROUND_ADVANCE_MONITOR_INTERVAL_SECONDS = 0.05
-_OCR_AFTER_ADVANCE_CAPTURE_DELAY_SECONDS = 0.15
-_OCR_AFTER_ADVANCE_SETTLE_POLL_SECONDS = 0.15
-_OCR_AFTER_ADVANCE_MAX_SETTLE_SECONDS = 2.0
+from .plugin_constants import (
+    _OCR_BACKEND_SELECTIONS,
+    _OCR_CAPTURE_BACKEND_SELECTIONS,
+    _BACKGROUND_BRIDGE_POLL_MIN_STALE_SECONDS,
+    _BRIDGE_TICK_INTERVAL_SECONDS,
+    _OCR_FOREGROUND_REFRESH_TTL_SECONDS,
+    _LATENCY_SAMPLE_LIMIT,
+    _LATENCY_MIN_SAMPLES_FOR_P95,
+    _OCR_POLL_P95_DEGRADE_THRESHOLD_SECONDS,
+    _OCR_FOREGROUND_ADVANCE_MONITOR_INTERVAL_SECONDS,
+    _OCR_AFTER_ADVANCE_CAPTURE_DELAY_SECONDS,
+    _OCR_AFTER_ADVANCE_SETTLE_POLL_SECONDS,
+    _OCR_AFTER_ADVANCE_MAX_SETTLE_SECONDS,
+)
 
 
 from .plugin_ocr_helpers import (
@@ -205,10 +205,90 @@ from .plugin_capture_profile_helpers import (
 
 
 from .plugin_config_service import GalgamePluginConfigService
+from .plugin_entries.galgame_get_status import _GalgameGetStatusMixin
+from .plugin_entries.galgame_install_textractor import _GalgameInstallTextractorMixin
+from .plugin_entries.galgame_download_rapidocr_models import _GalgameDownloadRapidocrModelsMixin
+from .plugin_entries.galgame_get_snapshot import _GalgameGetSnapshotMixin
+from .plugin_entries.galgame_get_history import _GalgameGetHistoryMixin
+from .plugin_entries.galgame_set_mode import _GalgameSetModeMixin
+from .plugin_entries.galgame_set_ocr_backend import _GalgameSetOcrBackendMixin
+from .plugin_entries.galgame_set_rapidocr_lang import _GalgameSetRapidocrLangMixin
+from .plugin_entries.galgame_set_ocr_timing import _GalgameSetOcrTimingMixin
+from .plugin_entries.galgame_set_llm_vision import _GalgameSetLlmVisionMixin
+from .plugin_entries.galgame_set_ocr_screen_templates import _GalgameSetOcrScreenTemplatesMixin
+from .plugin_entries.galgame_build_ocr_screen_template_draft import _GalgameBuildOcrScreenTemplateDraftMixin
+from .plugin_entries.galgame_validate_ocr_screen_templates import _GalgameValidateOcrScreenTemplatesMixin
+from .plugin_entries.galgame_get_ocr_screen_awareness_snapshot import _GalgameGetOcrScreenAwarenessSnapshotMixin
+from .plugin_entries.galgame_train_ocr_screen_awareness_model import _GalgameTrainOcrScreenAwarenessModelMixin
+from .plugin_entries.galgame_evaluate_ocr_screen_awareness_model import _GalgameEvaluateOcrScreenAwarenessModelMixin
+from .plugin_entries.galgame_bind_game import _GalgameBindGameMixin
+from .plugin_entries.galgame_set_ocr_capture_profile import _GalgameSetOcrCaptureProfileMixin
+from .plugin_entries.galgame_auto_recalibrate_ocr_dialogue_profile import _GalgameAutoRecalibrateOcrDialogueProfileMixin
+from .plugin_entries.galgame_apply_recommended_ocr_capture_profile import _GalgameApplyRecommendedOcrCaptureProfileMixin
+from .plugin_entries.galgame_rollback_ocr_capture_profile import _GalgameRollbackOcrCaptureProfileMixin
+from .plugin_entries.galgame_list_memory_reader_processes import _GalgameListMemoryReaderProcessesMixin
+from .plugin_entries.galgame_set_memory_reader_target import _GalgameSetMemoryReaderTargetMixin
+from .plugin_entries.galgame_list_ocr_windows import _GalgameListOcrWindowsMixin
+from .plugin_entries.galgame_set_ocr_window_target import _GalgameSetOcrWindowTargetMixin
+from .plugin_entries.galgame_open_ui import _GalgameOpenUiMixin
+from .plugin_entries.galgame_explain_line import _GalgameExplainLineMixin
+from .plugin_entries.galgame_summarize_scene import _GalgameSummarizeSceneMixin
+from .plugin_entries.galgame_suggest_choice import _GalgameSuggestChoiceMixin
+from .plugin_entries.galgame_agent_command import _GalgameAgentCommandMixin
+from .plugin_entries.galgame_continue_auto_advance import _GalgameContinueAutoAdvanceMixin
+from .plugin_entries.galgame_get_scene_context import _GalgameGetSceneContextMixin
+from .plugin_entries.galgame_get_story_so_far import _GalgameGetStorySoFarMixin
+from .plugin_entries.galgame_get_recent_lines import _GalgameGetRecentLinesMixin
+from .plugin_entries.galgame_get_push_history import _GalgameGetPushHistoryMixin
+from .plugin_entries.galgame_set_character_mode import _GalgameSetCharacterModeMixin
+from .plugin_entries.galgame_get_character_list import _GalgameGetCharacterListMixin
+from .plugin_entries.galgame_get_character_profile import _GalgameGetCharacterProfileMixin
+from .plugin_entries.galgame_import_character_data import _GalgameImportCharacterDataMixin
 
 
 @neko_plugin
-class GalgamePlugin(NekoPluginBase):
+class GalgamePlugin(
+    _GalgameGetStatusMixin,
+    _GalgameInstallTextractorMixin,
+    _GalgameDownloadRapidocrModelsMixin,
+    _GalgameGetSnapshotMixin,
+    _GalgameGetHistoryMixin,
+    _GalgameSetModeMixin,
+    _GalgameSetOcrBackendMixin,
+    _GalgameSetRapidocrLangMixin,
+    _GalgameSetOcrTimingMixin,
+    _GalgameSetLlmVisionMixin,
+    _GalgameSetOcrScreenTemplatesMixin,
+    _GalgameBuildOcrScreenTemplateDraftMixin,
+    _GalgameValidateOcrScreenTemplatesMixin,
+    _GalgameGetOcrScreenAwarenessSnapshotMixin,
+    _GalgameTrainOcrScreenAwarenessModelMixin,
+    _GalgameEvaluateOcrScreenAwarenessModelMixin,
+    _GalgameBindGameMixin,
+    _GalgameSetOcrCaptureProfileMixin,
+    _GalgameAutoRecalibrateOcrDialogueProfileMixin,
+    _GalgameApplyRecommendedOcrCaptureProfileMixin,
+    _GalgameRollbackOcrCaptureProfileMixin,
+    _GalgameListMemoryReaderProcessesMixin,
+    _GalgameSetMemoryReaderTargetMixin,
+    _GalgameListOcrWindowsMixin,
+    _GalgameSetOcrWindowTargetMixin,
+    _GalgameOpenUiMixin,
+    _GalgameExplainLineMixin,
+    _GalgameSummarizeSceneMixin,
+    _GalgameSuggestChoiceMixin,
+    _GalgameAgentCommandMixin,
+    _GalgameContinueAutoAdvanceMixin,
+    _GalgameGetSceneContextMixin,
+    _GalgameGetStorySoFarMixin,
+    _GalgameGetRecentLinesMixin,
+    _GalgameGetPushHistoryMixin,
+    _GalgameSetCharacterModeMixin,
+    _GalgameGetCharacterListMixin,
+    _GalgameGetCharacterProfileMixin,
+    _GalgameImportCharacterDataMixin,
+    NekoPluginBase,
+):
     def __init__(self, ctx):
         super().__init__(ctx)
         self.file_logger = self.enable_file_logging(log_level="INFO")
@@ -4488,1875 +4568,41 @@ class GalgamePlugin(NekoPluginBase):
             ),
         )
 
-    @plugin_entry(
-        id="galgame_get_status",
-        name=tr("entries.galgame_get_status.name", default='获取 galgame 插件状态'),
-        description=tr("entries.galgame_get_status.description", default='返回当前 bridge 连接状态、绑定游戏、最近错误与模式。'),
-        input_schema={"type": "object", "properties": {}},
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_status(self, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        return Ok(await self._build_status_payload_async())
 
-    @plugin_entry(
-        id="galgame_install_textractor",
-        name=tr("entries.galgame_install_textractor.name", default='安装 Textractor'),
-        description=tr("entries.galgame_install_textractor.description", default='检测并下载安装 TextractorCLI.exe，随后刷新 galgame_plugin 的桥接与读内存状态。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "force": {"type": "boolean", "default": False},
-            },
-        },
-        timeout=600.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_install_textractor(self, force: bool = False, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not self._textractor_install_lock.acquire(blocking=False):
-            return Err(SdkError(self._install_in_progress_message("Textractor")))
-        current_run_id = self._resolve_current_run_id(_)
-        progress_callback = self._resolve_install_progress_callback(current_run_id)
-        try:
-            install_result = await install_textractor(
-                logger=self.logger,
-                configured_path=self._cfg.memory_reader_textractor_path,
-                install_target_dir_raw=self._cfg.memory_reader_install_target_dir,
-                release_api_url=self._cfg.memory_reader_install_release_api_url,
-                timeout_seconds=self._cfg.memory_reader_install_timeout_seconds,
-                textractor_proxy=self._cfg.memory_reader_textractor_proxy,
-                force=bool(force),
-                task_id=current_run_id or None,
-                progress_callback=progress_callback,
-            )
-            clear_install_inspection_cache()
-            self._refresh_dependency_status()
-            await self._poll_bridge(force=True)
-            return Ok(
-                {
-                    "summary": str(install_result.get("summary") or self._install_ok_message("textractor", "Textractor")),
-                    "install_result": install_result,
-                    "status": await self._build_status_payload_async(),
-                }
-            )
-        except Exception as exc:
-            return Err(SdkError(self._format_install_entry_error("textractor", "Textractor", exc)))
-        finally:
-            self._textractor_install_lock.release()
 
     # NOTE: RapidOCR / DXcam runtime SDK install actions removed —
     # both packages are now bundled into the main program (see pyproject.toml
     # [dependency-groups] galgame). Run `uv sync --group galgame` for source
     # installs; packaged builds always include them.
 
-    @plugin_entry(
-        id="galgame_download_rapidocr_models",
-        name=tr("entries.galgame_download_rapidocr_models.name", default='下载 RapidOCR 模型'),
-        description=tr("entries.galgame_download_rapidocr_models.description", default='为当前 (lang_type, ocr_version) 选择优先从百度云下载缺失的 RapidOCR 模型文件到插件模型缓存目录，下载失败时自动回退至 ModelScope。bundled 默认（ch+PP-OCRv4）不需要下载。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "force": {"type": "boolean", "default": False},
-            },
-        },
-        timeout=600.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_download_rapidocr_models(self, force: bool = False, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not self._rapidocr_models_lock.acquire(blocking=False):
-            return Err(SdkError(self._install_in_progress_message("RapidOCR Models")))
-        current_run_id = self._resolve_current_run_id(_)
-        progress_callback = self._resolve_install_progress_callback(current_run_id)
-        try:
-            from .rapidocr_support import download_rapidocr_models
 
-            download_result = await download_rapidocr_models(
-                logger=self.logger,
-                install_target_dir_raw=self._cfg.rapidocr_install_target_dir,
-                ocr_version=self._cfg.rapidocr_ocr_version,
-                lang_type=self._cfg.rapidocr_lang_type,
-                timeout_seconds=float(self._cfg.ocr_reader_install_timeout_seconds or 180.0),
-                force=bool(force),
-                task_id=current_run_id or None,
-                progress_callback=progress_callback,
-                before_completed_callback=clear_install_inspection_cache,
-            )
-            clear_install_inspection_cache()
-            await self._poll_bridge(force=True)
-            downloaded = download_result.get("downloaded") or []
-            summary = (
-                f"RapidOCR models ready ({len(downloaded)} file(s) downloaded)"
-                if downloaded
-                else "RapidOCR models already present"
-            )
-            return Ok(
-                {
-                    "summary": summary,
-                    "download_result": download_result,
-                    "status": await self._build_status_payload_async(),
-                }
-            )
-        except Exception as exc:
-            return Err(SdkError(self._format_install_entry_error("rapidocr_models", "RapidOCR Models", exc)))
-        finally:
-            self._rapidocr_models_lock.release()
 
-    @plugin_entry(
-        id="galgame_get_snapshot",
-        name=tr("entries.galgame_get_snapshot.name", default='获取 galgame 快照'),
-        description=tr("entries.galgame_get_snapshot.description", default='返回当前游戏快照和 stale 状态。'),
-        input_schema={"type": "object", "properties": {}},
-        llm_result_fields=["snapshot"],
-    )
-    async def galgame_get_snapshot(self, **_):
-        state_snapshot = self._snapshot_state()
-        payload = build_snapshot_payload(SimpleNamespace(**state_snapshot))
-        return Ok(payload)
 
-    @plugin_entry(
-        id="galgame_get_history",
-        name=tr("entries.galgame_get_history.name", default='获取 galgame 历史'),
-        description=tr("entries.galgame_get_history.description", default='返回最近事件、稳定台词历史和选项历史。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "limit": {"type": "integer", "default": 50, "minimum": 1},
-                "include_events": {"type": "boolean", "default": True},
-            },
-        },
-        llm_result_fields=["stable_lines", "observed_lines", "choices"],
-    )
-    async def galgame_get_history(self, limit: int = 50, include_events: bool = True, **_):
-        with self._state_lock:
-            payload = build_history_payload(
-                self._state,
-                limit=max(1, int(limit)),
-                include_events=bool(include_events),
-            )
-        return Ok(payload)
 
-    @plugin_entry(
-        id="galgame_set_mode",
-        name=tr("entries.galgame_set_mode.name", default='设置 galgame 模式'),
-        description=tr("entries.galgame_set_mode.description", default='设置 silent / companion / choice_advisor 模式，并可选更新通知开关。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "mode": {"type": "string", "enum": sorted(MODES)},
-                "push_notifications": {"type": "boolean"},
-                "advance_speed": {"type": "string", "enum": sorted(ADVANCE_SPEEDS)},
-                "reader_mode": {"type": "string", "enum": sorted(READER_MODES)},
-            },
-            "required": ["mode"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_mode(
-        self,
-        mode: str,
-        push_notifications: bool | None = None,
-        advance_speed: str | None = None,
-        reader_mode: str | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if mode not in MODES:
-            return Err(SdkError(f"invalid galgame mode: {mode!r}"))
-        if advance_speed is not None and advance_speed not in ADVANCE_SPEEDS:
-            return Err(SdkError(f"invalid advance speed: {advance_speed!r}"))
-        try:
-            normalized_reader_mode = _normalize_reader_mode(reader_mode or self._cfg.reader_mode)
-        except ValueError as exc:
-            return Err(SdkError(str(exc)))
 
-        with self._state_lock:
-            current_mode = str(self._state.mode or "")
-            current_push_notifications = bool(self._state.push_notifications)
-            current_advance_speed = str(self._state.advance_speed or ADVANCE_SPEED_MEDIUM)
-        current_reader_mode = self._cfg.reader_mode
-        requested_push_notifications = (
-            bool(push_notifications)
-            if push_notifications is not None
-            else current_push_notifications
-        )
-        requested_advance_speed = (
-            str(advance_speed)
-            if advance_speed is not None
-            else current_advance_speed
-        )
-        if (
-            mode == current_mode
-            and requested_push_notifications == current_push_notifications
-            and requested_advance_speed == current_advance_speed
-            and normalized_reader_mode == current_reader_mode
-        ):
-            return Ok(
-                {
-                    "mode": current_mode,
-                    "push_notifications": current_push_notifications,
-                    "advance_speed": current_advance_speed,
-                    "reader_mode": current_reader_mode,
-                    "summary": (
-                        f"mode={current_mode} "
-                        f"push_notifications={current_push_notifications} "
-                        f"advance_speed={current_advance_speed} "
-                        f"reader_mode={current_reader_mode}"
-                    ),
-                    "skipped": True,
-                    "skip_reason": "already_applied",
-                }
-            )
 
-        with self._state_lock:
-            old_mode = str(self._state.mode or "")
-            old_push_notifications = bool(self._state.push_notifications)
-            old_advance_speed = str(self._state.advance_speed or ADVANCE_SPEED_MEDIUM)
-            old_active_data_source = str(self._state.active_data_source or "")
-            old_next_poll_at_monotonic = float(self._state.next_poll_at_monotonic or 0.0)
-            old_pending_ocr_advance_captures = int(self._pending_ocr_advance_captures or 0)
-            old_last_ocr_advance_capture_requested_at = float(
-                self._last_ocr_advance_capture_requested_at or 0.0
-            )
-            old_last_ocr_advance_capture_reason = str(
-                self._last_ocr_advance_capture_reason or ""
-            )
-        old_reader_mode = self._cfg.reader_mode
 
-        def _restore_mode_runtime_state() -> None:
-            self._cfg.reader_mode = old_reader_mode
-            with self._state_lock:
-                self._state.mode = old_mode
-                self._state.push_notifications = old_push_notifications
-                self._state.advance_speed = old_advance_speed
-                self._state.active_data_source = old_active_data_source
-                self._state.next_poll_at_monotonic = old_next_poll_at_monotonic
-                self._pending_ocr_advance_captures = old_pending_ocr_advance_captures
-                self._last_ocr_advance_capture_requested_at = (
-                    old_last_ocr_advance_capture_requested_at
-                )
-                self._last_ocr_advance_capture_reason = old_last_ocr_advance_capture_reason
-                self._state_dirty = True
-                self._cached_snapshot = None
-            for manager, label in (
-                (self._memory_reader_manager, "memory reader"),
-                (self._ocr_reader_manager, "OCR reader"),
-            ):
-                if manager is None:
-                    continue
-                try:
-                    manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame {} mode rollback update_config failed: {}",
-                        label,
-                        rollback_exc,
-                    )
 
-        # galgame_set_mode runs in the plugin's asyncio flow; simple config field
-        # assignment is atomic here and readers use getattr fallbacks.
-        self._cfg.reader_mode = normalized_reader_mode
-        try:
-            if self._memory_reader_manager is not None:
-                self._memory_reader_manager.update_config(self._cfg)
-            if self._ocr_reader_manager is not None:
-                self._ocr_reader_manager.update_config(self._cfg)
-        except Exception as exc:
-            _restore_mode_runtime_state()
-            return Err(SdkError(f"apply mode failed: {exc}"))
-        with self._state_lock:
-            self._state.mode = mode
-            if push_notifications is not None:
-                self._state.push_notifications = bool(push_notifications)
-            if advance_speed is not None:
-                self._state.advance_speed = advance_speed
-            if normalized_reader_mode == READER_MODE_MEMORY:
-                self._clear_pending_ocr_advance_captures_locked()
-            if not self._state.active_session_id:
-                self._state.active_data_source = _pending_data_source_for_reader_mode(
-                    normalized_reader_mode,
-                    memory_reader_allowed=normalized_reader_mode in {READER_MODE_AUTO, READER_MODE_MEMORY},
-                    ocr_reader_allowed=normalized_reader_mode in {READER_MODE_AUTO, READER_MODE_OCR},
-                    memory_reader_candidate_available=False,
-                )
-            self._state.next_poll_at_monotonic = 0.0
-            self._state_dirty = True
-            self._cached_snapshot = None
-            payload = {
-                "mode": self._state.mode,
-                "push_notifications": self._state.push_notifications,
-                "advance_speed": self._state.advance_speed,
-                "reader_mode": self._cfg.reader_mode,
-                "summary": (
-                    f"mode={self._state.mode} "
-                    f"push_notifications={self._state.push_notifications} "
-                    f"advance_speed={self._state.advance_speed} "
-                    f"reader_mode={self._cfg.reader_mode}"
-                ),
-            }
-            bound_game_id = self._state.bound_game_id
-            persist_push = self._state.push_notifications
-            persist_advance_speed = self._state.advance_speed
 
-        try:
-            self._config_service.persist_preferences(
-                bound_game_id=bound_game_id,
-                mode=mode,
-                push_notifications=persist_push,
-                advance_speed=persist_advance_speed,
-            )
-            self._config_service.persist_reader_mode(reader_mode=normalized_reader_mode)
-        except Exception as exc:
-            _restore_mode_runtime_state()
-            return Err(SdkError(f"persist mode failed: {exc}"))
-        await self._ensure_ocr_foreground_advance_monitor()
-        if (
-            mode_allows_agent_actuation(old_mode)
-            and not mode_allows_agent_actuation(mode)
-        ):
-            self.request_ocr_after_advance_capture(reason="mode_change_to_read_only")
-        self._start_background_bridge_poll()
 
-        # 进入 choice_advisor 时默认启用 OCR fast loop；离开时仅关闭自动开启的。
-        if mode == MODE_CHOICE_ADVISOR and old_mode != MODE_CHOICE_ADVISOR:
-            if not self._ocr_fast_loop_should_run():
-                if self._cfg is not None:
-                    self._cfg.ocr_reader_fast_loop_enabled = True
-                self._fast_loop_auto_enabled = True
-                self._start_ocr_fast_loop()
-        elif old_mode == MODE_CHOICE_ADVISOR and mode != MODE_CHOICE_ADVISOR:
-            if self._fast_loop_auto_enabled:
-                if self._cfg is not None:
-                    self._cfg.ocr_reader_fast_loop_enabled = False
-                self._fast_loop_auto_enabled = False
-                await self._cancel_ocr_fast_loop()
 
-        if self._game_agent is not None and not mode_allows_agent_actuation(mode):
-            try:
-                agent_payload = await self._game_agent.apply_mode_change(
-                    self._snapshot_state(include_private_context=True)
-                )
-                payload["agent"] = json_copy(agent_payload)
-            except Exception as exc:
-                payload["agent_warning"] = f"apply_mode_change failed: {exc}"
-        return Ok(payload)
 
-    @plugin_entry(
-        id="galgame_set_ocr_backend",
-        name=tr("entries.galgame_set_ocr_backend.name", default='设置 OCR / 截图后端'),
-        description=tr("entries.galgame_set_ocr_backend.description", default='切换 OCR 文本识别后端和截图后端。只影响 OCR 读取，不改变 Agent 点击安全策略。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "backend_selection": {
-                    "type": "string",
-                    "enum": sorted(_OCR_BACKEND_SELECTIONS),
-                },
-                "capture_backend": {
-                    "type": "string",
-                    "enum": sorted(_OCR_CAPTURE_BACKEND_SELECTIONS),
-                },
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_ocr_backend(
-        self,
-        backend_selection: str | None = None,
-        capture_backend: str | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        normalized_backend = str(backend_selection or "").strip().lower() or None
-        normalized_capture = str(capture_backend or "").strip().lower() or None
-        # Accept legacy "imagegrab" from external callers but normalize to "mss"
-        # before validation so the schema set can drop the deprecated value.
-        if normalized_capture == "imagegrab":
-            normalized_capture = "mss"
-        if normalized_backend is None and normalized_capture is None:
-            return Err(SdkError("backend_selection or capture_backend is required"))
-        if normalized_backend is not None and normalized_backend not in _OCR_BACKEND_SELECTIONS:
-            return Err(SdkError(f"invalid OCR backend: {backend_selection!r}"))
-        if normalized_capture is not None and normalized_capture not in _OCR_CAPTURE_BACKEND_SELECTIONS:
-            return Err(SdkError(f"invalid OCR capture backend: {capture_backend!r}"))
 
-        old_backend = self._cfg.ocr_reader_backend_selection
-        old_capture = self._cfg.ocr_reader_capture_backend
-        old_backend_explicit = bool(
-            getattr(self._cfg, "ocr_reader_backend_selection_explicit", False)
-        )
-        old_capture_explicit = bool(
-            getattr(self._cfg, "ocr_reader_capture_backend_explicit", False)
-        )
-        backend_changed = normalized_backend is not None and normalized_backend != old_backend
-        capture_changed = normalized_capture is not None and normalized_capture != old_capture
-        if normalized_backend is not None:
-            self._cfg.ocr_reader_backend_selection = normalized_backend
-            self._cfg.ocr_reader_backend_selection_explicit = True
-        if normalized_capture is not None:
-            self._cfg.ocr_reader_capture_backend = normalized_capture
-            self._cfg.ocr_reader_capture_backend_explicit = True
-        if self._ocr_reader_manager is not None:
-            try:
-                self._ocr_reader_manager.update_config(self._cfg)
-            except Exception as exc:
-                if normalized_backend is not None:
-                    self._cfg.ocr_reader_backend_selection = old_backend
-                    self._cfg.ocr_reader_backend_selection_explicit = old_backend_explicit
-                if normalized_capture is not None:
-                    self._cfg.ocr_reader_capture_backend = old_capture
-                    self._cfg.ocr_reader_capture_backend_explicit = old_capture_explicit
-                return Err(SdkError(f"apply OCR backend failed: {exc}"))
 
-        with self._state_lock:
-            self._state.next_poll_at_monotonic = 0.0
-            self._state_dirty = True
-            self._cached_snapshot = None
 
-        try:
-            self._config_service.persist_ocr_backend_selection(
-                backend_selection=normalized_backend,
-                capture_backend=normalized_capture,
-            )
-        except Exception as exc:
-            self._cfg.ocr_reader_backend_selection = old_backend
-            self._cfg.ocr_reader_capture_backend = old_capture
-            self._cfg.ocr_reader_backend_selection_explicit = old_backend_explicit
-            self._cfg.ocr_reader_capture_backend_explicit = old_capture_explicit
-            if self._ocr_reader_manager is not None:
-                try:
-                    self._ocr_reader_manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame OCR backend rollback update_config failed: {}",
-                        rollback_exc,
-                    )
-            return Err(SdkError(f"persist OCR backend failed: {exc}"))
 
-        if self._ocr_reader_manager is not None and (backend_changed or capture_changed):
-            reset_capture_runtime = getattr(
-                self._ocr_reader_manager,
-                "reset_capture_runtime_diagnostics",
-                None,
-            )
-            if callable(reset_capture_runtime):
-                try:
-                    reset_capture_runtime()
-                except Exception as exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame OCR backend switch diagnostic reset failed: {}",
-                        exc,
-                    )
 
-        self._start_background_bridge_poll()
-        payload = {
-            "backend_selection": self._cfg.ocr_reader_backend_selection,
-            "capture_backend": self._cfg.ocr_reader_capture_backend,
-            "summary": (
-                f"OCR backend={self._cfg.ocr_reader_backend_selection} "
-                f"capture_backend={self._cfg.ocr_reader_capture_backend}"
-            ),
-        }
-        return Ok(payload)
 
-    @plugin_entry(
-        id="galgame_set_rapidocr_lang",
-        name=tr("entries.galgame_set_rapidocr_lang.name", default='切换 OCR 识别语言'),
-        description=tr(
-            "entries.galgame_set_rapidocr_lang.description",
-            default='切换 RapidOCR 文字识别语言模型；手动切换语言后关闭自动检测。',
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "lang_type": {
-                    "type": "string",
-                    "enum": ["ch", "japan", "korean", "en"],
-                },
-                "auto_detect_lang": {
-                    "type": "boolean",
-                },
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_rapidocr_lang(
-        self,
-        lang_type: str | None = None,
-        auto_detect_lang: bool | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
 
-        normalized_lang = str(lang_type or "").strip().lower() or None
-        if normalized_lang is not None and normalized_lang not in {"ch", "japan", "korean", "en"}:
-            return Err(SdkError(f"invalid lang_type: {lang_type!r}"))
-        if normalized_lang is None and auto_detect_lang is None:
-            return Err(SdkError("lang_type or auto_detect_lang is required"))
 
-        old_lang = self._cfg.rapidocr_lang_type
-        old_auto = self._cfg.rapidocr_auto_detect_lang
-        old_last = self._cfg.rapidocr_auto_detect_last_lang
-        if normalized_lang is not None:
-            self._cfg.rapidocr_lang_type = normalized_lang
-            self._cfg.rapidocr_auto_detect_last_lang = normalized_lang
-            self._cfg.rapidocr_auto_detect_lang = False
-        if normalized_lang is None and auto_detect_lang is not None:
-            self._cfg.rapidocr_auto_detect_lang = bool(auto_detect_lang)
 
-        if self._ocr_reader_manager is not None:
-            try:
-                self._ocr_reader_manager.update_config(self._cfg)
-            except Exception as exc:
-                self._cfg.rapidocr_lang_type = old_lang
-                self._cfg.rapidocr_auto_detect_lang = old_auto
-                self._cfg.rapidocr_auto_detect_last_lang = old_last
-                return Err(SdkError(f"apply rapidocr lang failed: {exc}"))
 
-        with self._state_lock:
-            self._state.next_poll_at_monotonic = 0.0
-            self._state_dirty = True
-            self._cached_snapshot = None
 
-        try:
-            self._config_service.persist_rapidocr_lang(
-                lang_type=normalized_lang,
-                auto_detect_lang=(
-                    bool(auto_detect_lang)
-                    if normalized_lang is None and auto_detect_lang is not None
-                    else (False if normalized_lang is not None else None)
-                ),
-                auto_detect_last_lang=normalized_lang,
-            )
-        except Exception as exc:
-            self._cfg.rapidocr_lang_type = old_lang
-            self._cfg.rapidocr_auto_detect_lang = old_auto
-            self._cfg.rapidocr_auto_detect_last_lang = old_last
-            if self._ocr_reader_manager is not None:
-                try:
-                    self._ocr_reader_manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame rapidocr lang rollback update_config failed: {}",
-                        rollback_exc,
-                    )
-            return Err(SdkError(f"persist rapidocr lang failed: {exc}"))
 
-        self._refresh_dependency_status()
-        self._start_background_bridge_poll()
-        return Ok({
-            "lang_type": self._cfg.rapidocr_lang_type,
-            "auto_detect_lang": self._cfg.rapidocr_auto_detect_lang,
-            "summary": (
-                f"RapidOCR lang={self._cfg.rapidocr_lang_type} "
-                f"auto_detect={'on' if self._cfg.rapidocr_auto_detect_lang else 'off'}"
-            ),
-        })
 
-    @plugin_entry(
-        id="galgame_set_ocr_timing",
-        name=tr("entries.galgame_set_ocr_timing.name", default='设置 OCR 识别时机'),
-        description=tr("entries.galgame_set_ocr_timing.description", default='设置 OCR Reader 触发模式与轮询间隔；DXcam 截图后端会随 OCR 触发。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "poll_interval_seconds": {
-                    "type": "number",
-                    "minimum": 0.5,
-                    "maximum": 10.0,
-                },
-                "trigger_mode": {
-                    "type": "string",
-                    "enum": ["interval", "after_advance"],
-                    "default": "interval",
-                },
-                "fast_loop_enabled": {"type": "boolean"},
-            },
-            "required": ["poll_interval_seconds"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_ocr_timing(
-        self,
-        poll_interval_seconds: float,
-        trigger_mode: str | None = None,
-        fast_loop_enabled: bool | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        try:
-            normalized_interval = float(poll_interval_seconds)
-        except (TypeError, ValueError):
-            return Err(SdkError("poll_interval_seconds must be a number"))
-        if normalized_interval < 0.5 or normalized_interval > 10.0:
-            return Err(SdkError("poll_interval_seconds must be between 0.5 and 10.0"))
-        try:
-            normalized_trigger_mode = _normalize_ocr_trigger_mode(
-                trigger_mode or self._cfg.ocr_reader_trigger_mode
-            )
-        except ValueError as exc:
-            return Err(SdkError(str(exc)))
 
-        old_interval = self._cfg.ocr_reader_poll_interval_seconds
-        old_trigger_mode = self._cfg.ocr_reader_trigger_mode
-        old_fast_loop = self._cfg.ocr_reader_fast_loop_enabled
-        old_fast_loop_auto_enabled = self._fast_loop_auto_enabled
-        if fast_loop_enabled is not None:
-            self._fast_loop_auto_enabled = False
-        self._cfg.ocr_reader_poll_interval_seconds = normalized_interval
-        self._cfg.ocr_reader_trigger_mode = normalized_trigger_mode
-        self._cfg.ocr_reader_fast_loop_enabled = (
-            bool(fast_loop_enabled)
-            if fast_loop_enabled is not None
-            else old_fast_loop
-        )
-        if self._ocr_reader_manager is not None:
-            try:
-                self._ocr_reader_manager.update_config(self._cfg)
-            except Exception as exc:
-                self._cfg.ocr_reader_poll_interval_seconds = old_interval
-                self._cfg.ocr_reader_trigger_mode = old_trigger_mode
-                self._cfg.ocr_reader_fast_loop_enabled = old_fast_loop
-                self._fast_loop_auto_enabled = old_fast_loop_auto_enabled
-                return Err(SdkError(f"apply OCR timing failed: {exc}"))
 
-        with self._state_lock:
-            self._state.next_poll_at_monotonic = 0.0
-            self._state_dirty = True
-            self._cached_snapshot = None
 
-        try:
-            self._config_service.persist_ocr_timing(
-                poll_interval_seconds=normalized_interval,
-                trigger_mode=normalized_trigger_mode,
-                fast_loop_enabled=(
-                    fast_loop_enabled
-                    if fast_loop_enabled is not None
-                    else old_fast_loop
-                ),
-            )
-        except Exception as exc:
-            self._cfg.ocr_reader_poll_interval_seconds = old_interval
-            self._cfg.ocr_reader_trigger_mode = old_trigger_mode
-            self._cfg.ocr_reader_fast_loop_enabled = old_fast_loop
-            self._fast_loop_auto_enabled = old_fast_loop_auto_enabled
-            if self._ocr_reader_manager is not None:
-                try:
-                    self._ocr_reader_manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame OCR timing rollback update_config failed: {}",
-                        rollback_exc,
-                    )
-            return Err(SdkError(f"persist OCR timing failed: {exc}"))
-
-        if normalized_trigger_mode != OCR_TRIGGER_MODE_AFTER_ADVANCE:
-            self._clear_pending_ocr_advance_captures()
-        if fast_loop_enabled is not None:
-            try:
-                if bool(fast_loop_enabled) and not old_fast_loop:
-                    self._start_ocr_fast_loop()
-                elif not bool(fast_loop_enabled) and old_fast_loop:
-                    await self._cancel_ocr_fast_loop()
-            except Exception as exc:
-                self._cfg.ocr_reader_poll_interval_seconds = old_interval
-                self._cfg.ocr_reader_trigger_mode = old_trigger_mode
-                self._cfg.ocr_reader_fast_loop_enabled = old_fast_loop
-                self._fast_loop_auto_enabled = old_fast_loop_auto_enabled
-                if self._ocr_reader_manager is not None:
-                    try:
-                        self._ocr_reader_manager.update_config(self._cfg)
-                    except Exception as rollback_exc:
-                        _log_plugin_noncritical(
-                            self.logger,
-                            "warning",
-                            "galgame OCR timing rollback update_config failed: {}",
-                            rollback_exc,
-                        )
-                try:
-                    self._config_service.persist_ocr_timing(
-                        poll_interval_seconds=old_interval,
-                        trigger_mode=old_trigger_mode,
-                        fast_loop_enabled=old_fast_loop,
-                    )
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame OCR fast loop rollback persist failed: {}",
-                        rollback_exc,
-                    )
-                return Err(SdkError(f"apply fast_loop_enabled failed: {exc}"))
-        await self._ensure_ocr_foreground_advance_monitor()
-        self._start_background_bridge_poll()
-        trigger_mode_label = (
-            "点击对白后识别"
-            if self._cfg.ocr_reader_trigger_mode == OCR_TRIGGER_MODE_AFTER_ADVANCE
-            else "按间隔识别"
-        )
-        payload = {
-            "poll_interval_seconds": self._cfg.ocr_reader_poll_interval_seconds,
-            "trigger_mode": self._cfg.ocr_reader_trigger_mode,
-            "fast_loop_enabled": self._cfg.ocr_reader_fast_loop_enabled,
-            "summary": (
-                f"OCR/DXcam {trigger_mode_label}；间隔="
-                f"{self._cfg.ocr_reader_poll_interval_seconds:.1f}s；"
-                f"Fast Loop={'开启' if self._cfg.ocr_reader_fast_loop_enabled else '关闭'}"
-            ),
-        }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_set_llm_vision",
-        name=tr("entries.galgame_set_llm_vision.name", default='设置 LLM 视觉辅助'),
-        description=tr("entries.galgame_set_llm_vision.description", default='切换 OCR Agent 低置信度场景的截图直传，并设置图片最大边长。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "vision_enabled": {"type": "boolean"},
-                "vision_max_image_px": {
-                    "type": "integer",
-                    "minimum": 64,
-                    "maximum": 2048,
-                    "default": 768,
-                },
-            },
-            "required": ["vision_enabled"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_llm_vision(
-        self,
-        vision_enabled: bool,
-        vision_max_image_px: int | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        try:
-            normalized_max_px = int(
-                vision_max_image_px
-                if vision_max_image_px is not None
-                else self._cfg.llm_vision_max_image_px
-            )
-        except (TypeError, ValueError):
-            return Err(SdkError("vision_max_image_px must be an integer"))
-        if normalized_max_px < 64 or normalized_max_px > 2048:
-            return Err(SdkError("vision_max_image_px must be between 64 and 2048"))
-
-        old_enabled = self._cfg.llm_vision_enabled
-        old_max_px = self._cfg.llm_vision_max_image_px
-        self._cfg.llm_vision_enabled = bool(vision_enabled)
-        self._cfg.llm_vision_max_image_px = normalized_max_px
-        if self._ocr_reader_manager is not None:
-            try:
-                self._ocr_reader_manager.update_config(self._cfg)
-            except Exception as exc:
-                self._cfg.llm_vision_enabled = old_enabled
-                self._cfg.llm_vision_max_image_px = old_max_px
-                return Err(SdkError(f"apply LLM vision failed: {exc}"))
-
-        with self._state_lock:
-            self._state_dirty = True
-            self._cached_snapshot = None
-
-        try:
-            self._config_service.persist_llm_vision(
-                vision_enabled=self._cfg.llm_vision_enabled,
-                vision_max_image_px=self._cfg.llm_vision_max_image_px,
-            )
-        except Exception as exc:
-            self._cfg.llm_vision_enabled = old_enabled
-            self._cfg.llm_vision_max_image_px = old_max_px
-            if self._ocr_reader_manager is not None:
-                try:
-                    self._ocr_reader_manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame LLM vision rollback update_config failed: {}",
-                        rollback_exc,
-                    )
-            return Err(SdkError(f"persist LLM vision failed: {exc}"))
-
-        payload = {
-            "vision_enabled": self._cfg.llm_vision_enabled,
-            "vision_max_image_px": self._cfg.llm_vision_max_image_px,
-            "summary": (
-                f"LLM vision enabled={self._cfg.llm_vision_enabled} "
-                f"max_image_px={self._cfg.llm_vision_max_image_px}"
-            ),
-        }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_set_ocr_screen_templates",
-        name=tr("entries.galgame_set_ocr_screen_templates.name", default='设置 OCR 屏幕模板'),
-        description=tr("entries.galgame_set_ocr_screen_templates.description", default='保存 OCR 屏幕分类模板；模板仅影响 OCR Reader，不影响 Bridge SDK / Memory Reader。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "screen_templates": {
-                    "type": "array",
-                    "items": {"type": "object"},
-                    "default": [],
-                },
-            },
-            "required": ["screen_templates"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_ocr_screen_templates(
-        self,
-        screen_templates: list[dict[str, Any]] | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not isinstance(screen_templates, list):
-            return Err(SdkError("screen_templates must be an array"))
-        sanitized = build_config(
-            {"ocr_reader": {"screen_templates": screen_templates}}
-        ).ocr_reader_screen_templates
-        old_templates = json_copy(self._cfg.ocr_reader_screen_templates)
-        self._cfg.ocr_reader_screen_templates = json_copy(sanitized)
-        if self._ocr_reader_manager is not None:
-            try:
-                self._ocr_reader_manager.update_config(self._cfg)
-            except Exception as exc:
-                self._cfg.ocr_reader_screen_templates = old_templates
-                return Err(SdkError(f"apply OCR screen templates failed: {exc}"))
-
-        with self._state_lock:
-            self._state_dirty = True
-            self._cached_snapshot = None
-
-        try:
-            self._config_service.persist_ocr_screen_templates(
-                self._cfg.ocr_reader_screen_templates
-            )
-        except Exception as exc:
-            self._cfg.ocr_reader_screen_templates = old_templates
-            if self._ocr_reader_manager is not None:
-                try:
-                    self._ocr_reader_manager.update_config(self._cfg)
-                except Exception as rollback_exc:
-                    _log_plugin_noncritical(
-                        self.logger,
-                        "warning",
-                        "galgame OCR screen template rollback update_config failed: {}",
-                        rollback_exc,
-                    )
-            return Err(SdkError(f"persist OCR screen templates failed: {exc}"))
-
-        payload = {
-            "screen_templates": json_copy(self._cfg.ocr_reader_screen_templates),
-            "summary": f"OCR screen templates={len(self._cfg.ocr_reader_screen_templates)}",
-        }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_build_ocr_screen_template_draft",
-        name=tr("entries.galgame_build_ocr_screen_template_draft.name", default='生成 OCR 屏幕模板草稿'),
-        description=tr("entries.galgame_build_ocr_screen_template_draft.description", default='根据当前 OCR 运行时、窗口信息和最近识别文本生成可编辑的屏幕模板草稿。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "stage": {
-                    "type": "string",
-                    "enum": sorted(OCR_CAPTURE_PROFILE_STAGES),
-                },
-                "region": {"type": "object"},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_build_ocr_screen_template_draft(
-        self,
-        stage: str | None = None,
-        region: dict[str, Any] | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        try:
-            payload = self._build_ocr_screen_template_draft_payload(
-                stage=stage,
-                region=region,
-            )
-        except Exception as exc:
-            return Err(SdkError(f"build OCR screen template draft failed: {exc}"))
-        payload["summary"] = (
-            f"OCR screen template draft stage={payload['template'].get('stage')} "
-            f"id={payload['template'].get('id')}"
-        )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_validate_ocr_screen_templates",
-        name=tr("entries.galgame_validate_ocr_screen_templates.name", default='验证 OCR 屏幕模板'),
-        description=tr("entries.galgame_validate_ocr_screen_templates.description", default='用当前 OCR 运行时和最近文本回放验证屏幕模板命中结果。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "screen_templates": {
-                    "type": "array",
-                    "items": {"type": "object"},
-                    "default": [],
-                },
-            },
-            "required": ["screen_templates"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_validate_ocr_screen_templates(
-        self,
-        screen_templates: list[dict[str, Any]] | None = None,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not isinstance(screen_templates, list):
-            return Err(SdkError("screen_templates must be an array"))
-        try:
-            payload = self._current_ocr_screen_template_validation_payload(screen_templates)
-        except Exception as exc:
-            return Err(SdkError(f"validate OCR screen templates failed: {exc}"))
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_get_ocr_screen_awareness_snapshot",
-        name=tr("entries.galgame_get_ocr_screen_awareness_snapshot.name", default='获取 OCR 屏幕感知截图'),
-        description=tr("entries.galgame_get_ocr_screen_awareness_snapshot.description", default='返回最近一次 OCR 屏幕感知截图；仅在 Vision 显式开启且短期缓存未过期时可用。'),
-        input_schema={"type": "object", "properties": {}},
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_ocr_screen_awareness_snapshot(self, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        snapshot = self.latest_ocr_vision_snapshot()
-        if not isinstance(snapshot, dict) or not snapshot.get("vision_image_base64"):
-            return Err(SdkError("no OCR screen awareness snapshot is available; enable Vision and wait for a full-frame OCR capture"))
-        payload = {
-            "snapshot": snapshot,
-            "summary": (
-                f"OCR screen awareness snapshot "
-                f"{int(snapshot.get('width') or 0)}x{int(snapshot.get('height') or 0)}"
-            ),
-        }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_train_ocr_screen_awareness_model",
-        name=tr("entries.galgame_train_ocr_screen_awareness_model.name", default='训练 OCR 屏幕感知模型'),
-        description=tr("entries.galgame_train_ocr_screen_awareness_model.description", default='从已标注 JSONL 样本训练轻量原型分类器，并导出可部署 JSON 模型。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "sample_path": {"type": "string", "default": ""},
-                "output_path": {"type": "string", "default": "screen_awareness_model.json"},
-                "allow_rule_labels": {"type": "boolean", "default": False},
-                "validation_ratio": {"type": "number", "default": 0.2},
-                "min_samples_per_stage": {"type": "integer", "default": 2},
-                "min_confidence": {"type": "number", "default": 0.55},
-            },
-        },
-        timeout=120.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_train_ocr_screen_awareness_model(
-        self,
-        sample_path: str = "",
-        output_path: str = "screen_awareness_model.json",
-        allow_rule_labels: bool = False,
-        validation_ratio: float = 0.2,
-        min_samples_per_stage: int = 2,
-        min_confidence: float = 0.55,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        try:
-            resolved_samples = self._resolve_screen_awareness_data_path(
-                sample_path,
-                default_filename="samples.jsonl",
-            )
-            resolved_output = self._resolve_screen_awareness_data_path(
-                output_path,
-                default_filename="screen_awareness_model.json",
-            )
-            result = await asyncio.to_thread(
-                train_screen_awareness_model,
-                resolved_samples,
-                resolved_output,
-                allow_rule_labels=bool(allow_rule_labels),
-                validation_ratio=float(validation_ratio),
-                min_samples_per_stage=int(min_samples_per_stage),
-                min_confidence=float(min_confidence),
-            )
-        except Exception as exc:
-            return Err(SdkError(f"train OCR screen awareness model failed: {exc}"))
-        payload = {
-            "output_path": str(result.get("output_path") or resolved_output),
-            "evaluation": json_copy(result.get("evaluation") or {}),
-            "model": json_copy(result.get("model") or {}),
-            "summary": str(result.get("summary") or "OCR screen awareness model trained"),
-        }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_evaluate_ocr_screen_awareness_model",
-        name=tr("entries.galgame_evaluate_ocr_screen_awareness_model.name", default='评估 OCR 屏幕感知模型'),
-        description=tr("entries.galgame_evaluate_ocr_screen_awareness_model.description", default='用已标注 JSONL 样本评估轻量屏幕感知模型，并可输出评估报告。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "sample_path": {"type": "string", "default": ""},
-                "model_path": {"type": "string", "default": "screen_awareness_model.json"},
-                "report_path": {"type": "string", "default": ""},
-                "allow_rule_labels": {"type": "boolean", "default": False},
-                "min_confidence": {"type": "number", "default": 0.55},
-            },
-        },
-        timeout=120.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_evaluate_ocr_screen_awareness_model(
-        self,
-        sample_path: str = "",
-        model_path: str = "screen_awareness_model.json",
-        report_path: str = "",
-        allow_rule_labels: bool = False,
-        min_confidence: float = 0.55,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        try:
-            resolved_samples = self._resolve_screen_awareness_data_path(
-                sample_path,
-                default_filename="samples.jsonl",
-            )
-            resolved_model = self._resolve_screen_awareness_data_path(
-                model_path,
-                default_filename="screen_awareness_model.json",
-            )
-            resolved_report = (
-                self._resolve_screen_awareness_data_path(
-                    report_path,
-                    default_filename="screen_awareness_evaluation.json",
-                )
-                if str(report_path or "").strip()
-                else None
-            )
-            result = await asyncio.to_thread(
-                evaluate_screen_awareness_model,
-                resolved_samples,
-                resolved_model,
-                allow_rule_labels=bool(allow_rule_labels),
-                min_confidence=float(min_confidence),
-                report_path=resolved_report,
-            )
-        except Exception as exc:
-            return Err(SdkError(f"evaluate OCR screen awareness model failed: {exc}"))
-        return Ok(json_copy(result))
-
-    @plugin_entry(
-        id="galgame_bind_game",
-        name=tr("entries.galgame_bind_game.name", default='绑定 galgame 游戏'),
-        description=tr("entries.galgame_bind_game.description", default='绑定指定 game_id；传空字符串清除手动绑定并恢复自动选择。'),
-        input_schema={
-            "type": "object",
-            "properties": {"game_id": {"type": "string", "default": ""}},
-            "required": ["game_id"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_bind_game(self, game_id: str, **_):
-        normalized = game_id.strip()
-        with self._state_lock:
-            available_game_ids = list(self._state.available_game_ids)
-        if normalized and normalized not in available_game_ids:
-            return Err(SdkError(f"unknown game_id: {normalized!r}"))
-
-        with self._state_lock:
-            self._state.bound_game_id = normalized
-            self._state_dirty = True
-            self._cached_snapshot = None
-            bound_game_id = self._state.bound_game_id
-            mode = self._state.mode
-            push_notifications = self._state.push_notifications
-            advance_speed = self._state.advance_speed
-        self._clear_character_profiles()
-
-        try:
-            self._config_service.persist_preferences(
-                bound_game_id=bound_game_id,
-                mode=mode,
-                push_notifications=push_notifications,
-                advance_speed=advance_speed,
-            )
-        except Exception as exc:
-            return Err(SdkError(f"persist binding failed: {exc}"))
-
-        self._load_character_profiles_for_current_context(force=True)
-        await self._poll_bridge(force=True)
-        with self._state_lock:
-            payload = {
-                "bound_game_id": self._state.bound_game_id,
-                "active_session_id": self._state.active_session_id,
-                "summary": f"bound_game_id={self._state.bound_game_id or '(auto)'} active_session_id={self._state.active_session_id}",
-            }
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_set_ocr_capture_profile",
-        name=tr("entries.galgame_set_ocr_capture_profile.name", default='设置 OCR 截图校准'),
-        description=tr("entries.galgame_set_ocr_capture_profile.description", default='按进程名保存或清除 OCR Reader 的截图裁剪配置。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "process_name": {"type": "string", "default": ""},
-                "stage": {
-                    "type": "string",
-                    "enum": sorted(OCR_CAPTURE_PROFILE_STAGES),
-                    "default": OCR_CAPTURE_PROFILE_STAGE_DEFAULT,
-                },
-                "save_scope": {
-                    "type": "string",
-                    "enum": sorted(OCR_CAPTURE_PROFILE_SAVE_SCOPES),
-                },
-                "left_inset_ratio": {"type": "number", "default": 0.05},
-                "right_inset_ratio": {"type": "number", "default": 0.05},
-                "top_ratio": {"type": "number", "default": 0.3},
-                "bottom_inset_ratio": {"type": "number", "default": 0.3},
-                "clear": {"type": "boolean", "default": False},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_ocr_capture_profile(
-        self,
-        process_name: str = "",
-        stage: str = OCR_CAPTURE_PROFILE_STAGE_DEFAULT,
-        left_inset_ratio: float = 0.05,
-        right_inset_ratio: float = 0.05,
-        top_ratio: float = 0.3,
-        bottom_inset_ratio: float = 0.3,
-        save_scope: str | None = None,
-        clear: bool = False,
-        **_,
-    ):
-        def _parse_ratio(name: str, value: float) -> float:
-            if isinstance(value, bool):
-                raise ValueError(f"{name} must be a number")
-            try:
-                parsed = float(value)
-            except (TypeError, ValueError) as exc:
-                raise ValueError(f"{name} must be a number") from exc
-            if parsed < 0.0 or parsed >= 1.0:
-                raise ValueError(f"{name} must be >= 0.0 and < 1.0")
-            return parsed
-
-        with self._state_lock:
-            runtime_process_name = str(
-                (self._state.ocr_reader_runtime or {}).get("process_name") or ""
-            ).strip()
-        normalized_process_name = str(process_name or "").strip() or runtime_process_name
-        if not normalized_process_name:
-            return Err(SdkError("process_name is required"))
-
-        if clear:
-            normalized_profile: dict[str, float] | None = None
-        else:
-            try:
-                normalized_profile = {
-                    "left_inset_ratio": _parse_ratio("left_inset_ratio", left_inset_ratio),
-                    "right_inset_ratio": _parse_ratio("right_inset_ratio", right_inset_ratio),
-                    "top_ratio": _parse_ratio("top_ratio", top_ratio),
-                    "bottom_inset_ratio": _parse_ratio(
-                        "bottom_inset_ratio",
-                        bottom_inset_ratio,
-                    ),
-                }
-            except ValueError as exc:
-                return Err(SdkError(str(exc)))
-            if (
-                normalized_profile["left_inset_ratio"]
-                + normalized_profile["right_inset_ratio"]
-            ) >= 1.0:
-                return Err(SdkError("left_inset_ratio + right_inset_ratio must be < 1.0"))
-            if (
-                normalized_profile["top_ratio"]
-                + normalized_profile["bottom_inset_ratio"]
-            ) >= 1.0:
-                return Err(SdkError("top_ratio + bottom_inset_ratio must be < 1.0"))
-        try:
-            payload = await self._save_ocr_capture_profile_payload(
-                process_name=normalized_process_name,
-                stage=stage,
-                capture_profile=normalized_profile,
-                clear=bool(clear),
-                save_scope=save_scope,
-            )
-        except ValueError as exc:
-            return Err(SdkError(str(exc)))
-        except Exception as exc:
-            return Err(SdkError(f"persist OCR capture profile failed: {exc}"))
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_auto_recalibrate_ocr_dialogue_profile",
-        name=tr("entries.galgame_auto_recalibrate_ocr_dialogue_profile.name", default='自动重新校准 OCR 对白区'),
-        description=tr("entries.galgame_auto_recalibrate_ocr_dialogue_profile.description", default='对当前已附着 OCR 目标窗口自动重校准对白区，并保存到当前窗口分辨率。'),
-        input_schema={"type": "object", "properties": {}},
-        timeout=120.0,
-        llm_result_fields=["summary", "sample_text"],
-    )
-    async def galgame_auto_recalibrate_ocr_dialogue_profile(self, **_):
-        if self._ocr_reader_manager is None:
-            return Err(SdkError("ocr_reader manager is not initialized"))
-        try:
-            recalibrated = await asyncio.to_thread(
-                self._ocr_reader_manager.auto_recalibrate_dialogue_profile
-            )
-            payload = await self._save_ocr_capture_profile_payload(
-                process_name=str(recalibrated.get("process_name") or ""),
-                stage=OCR_CAPTURE_PROFILE_STAGE_DIALOGUE,
-                capture_profile=dict(recalibrated.get("capture_profile") or {}),
-                clear=False,
-                save_scope=OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET,
-                width=int(recalibrated.get("window_width") or 0),
-                height=int(recalibrated.get("window_height") or 0),
-            )
-        except ValueError as exc:
-            return Err(SdkError(str(exc)))
-        except Exception as exc:
-            return Err(SdkError(f"auto recalibrate OCR dialogue profile failed: {exc}"))
-        payload.update(
-            {
-                "sample_text": str(recalibrated.get("sample_text") or ""),
-                "save_scope": OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET,
-                "bucket_key": str(recalibrated.get("bucket_key") or payload.get("bucket_key") or ""),
-                "window_width": int(
-                    recalibrated.get("window_width") or payload.get("window_width") or 0
-                ),
-                "window_height": int(
-                    recalibrated.get("window_height") or payload.get("window_height") or 0
-                ),
-                "summary": str(recalibrated.get("summary") or payload.get("summary") or ""),
-            }
-        )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_apply_recommended_ocr_capture_profile",
-        name=tr("entries.galgame_apply_recommended_ocr_capture_profile.name", default='应用推荐 OCR 截图校准'),
-        description=tr("entries.galgame_apply_recommended_ocr_capture_profile.description", default='在用户确认后应用当前 OCR Reader 推荐的截图 profile，并记录回滚点。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "confirm": {"type": "boolean", "default": False},
-                "enable_auto_apply": {"type": "boolean", "default": False},
-                "allow_manual_override": {"type": "boolean", "default": False},
-            },
-            "required": ["confirm"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_apply_recommended_ocr_capture_profile(
-        self,
-        confirm: bool = False,
-        enable_auto_apply: bool = False,
-        allow_manual_override: bool = False,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not bool(confirm):
-            return Err(SdkError("confirm=true is required before applying a recommended OCR profile"))
-        with self._state_lock:
-            runtime = json_copy(self._state.ocr_reader_runtime)
-        self._ocr_capture_profile_auto_apply_enabled = bool(enable_auto_apply)
-        try:
-            payload = await self._apply_recommended_ocr_capture_profile_payload(
-                runtime,
-                allow_manual_override=bool(allow_manual_override),
-                reason="manual_apply_recommended_capture_profile",
-            )
-        except ValueError as exc:
-            return Err(SdkError(str(exc)))
-        except Exception as exc:
-            return Err(SdkError(f"apply recommended OCR capture profile failed: {exc}"))
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_rollback_ocr_capture_profile",
-        name=tr("entries.galgame_rollback_ocr_capture_profile.name", default='回滚 OCR 推荐截图校准'),
-        description=tr("entries.galgame_rollback_ocr_capture_profile.description", default='回滚最近一次由推荐 profile 应用产生的 OCR 截图校准。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "confirm": {"type": "boolean", "default": False},
-            },
-            "required": ["confirm"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_rollback_ocr_capture_profile(
-        self,
-        confirm: bool = False,
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        if not bool(confirm):
-            return Err(SdkError("confirm=true is required before rolling back OCR profile"))
-        try:
-            payload = await self._rollback_pending_ocr_capture_profile(
-                reason="manual_rollback_recommended_capture_profile"
-            )
-        except Exception as exc:
-            return Err(SdkError(f"rollback OCR capture profile failed: {exc}"))
-        if payload is None:
-            return Err(SdkError("no pending recommended OCR capture profile rollback"))
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_list_memory_reader_processes",
-        name=tr("entries.galgame_list_memory_reader_processes.name", default='列出 Memory Reader 候选进程'),
-        description=tr("entries.galgame_list_memory_reader_processes.description", default='返回 Memory Reader 可选进程，包含 exe 路径、检测到的引擎和识别原因。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "include_unknown": {"type": "boolean", "default": True},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_list_memory_reader_processes(
-        self,
-        include_unknown: bool = True,
-        **_,
-    ):
-        if self._memory_reader_manager is None:
-            return Err(SdkError("memory_reader manager is not initialized"))
-        try:
-            payload = await asyncio.to_thread(
-                self._memory_reader_manager.list_processes_snapshot,
-                include_unknown=bool(include_unknown),
-            )
-        except Exception as exc:
-            return Err(SdkError(f"list Memory Reader processes failed: {exc}"))
-        payload["summary"] = (
-            f"processes={int(payload.get('candidate_count') or 0)} "
-            f"mode={payload.get('target_selection_mode') or 'auto'}"
-        )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_set_memory_reader_target",
-        name=tr("entries.galgame_set_memory_reader_target.name", default='设置 Memory Reader 目标进程'),
-        description=tr("entries.galgame_set_memory_reader_target.description", default='锁定或清除 Memory Reader 的手动进程目标。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "process_key": {"type": "string", "default": ""},
-                "pid": {"type": "integer", "default": 0},
-                "exe_path": {"type": "string", "default": ""},
-                "process_name": {"type": "string", "default": ""},
-                "clear": {"type": "boolean", "default": False},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_memory_reader_target(
-        self,
-        process_key: str = "",
-        pid: int = 0,
-        exe_path: str = "",
-        process_name: str = "",
-        clear: bool = False,
-        **_,
-    ):
-        if self._memory_reader_manager is None:
-            return Err(SdkError("memory_reader manager is not initialized"))
-
-        if clear:
-            target_payload = {
-                "mode": "auto",
-                "process_key": "",
-                "process_name": "",
-                "exe_path": "",
-                "pid": 0,
-                "engine": "",
-                "detected_engine": "",
-                "detection_reason": "",
-                "create_time": 0.0,
-                "selected_at": "",
-            }
-            summary = "Memory Reader process target cleared; using auto detection"
-        else:
-            try:
-                target_payload = await asyncio.to_thread(
-                    self._memory_reader_manager.resolve_manual_process_target,
-                    process_key=process_key,
-                    pid=pid,
-                    exe_path=exe_path,
-                    process_name=process_name,
-                )
-            except ValueError as exc:
-                return Err(SdkError(str(exc)))
-            except Exception as exc:
-                return Err(SdkError(f"resolve Memory Reader process target failed: {exc}"))
-            summary = (
-                f"Memory Reader target locked to {target_payload.get('process_name') or '(unknown)'}"
-            )
-
-        try:
-            self._persist.persist_memory_reader_target(target_payload)
-        except Exception as exc:
-            return Err(SdkError(f"persist Memory Reader target failed: {exc}"))
-
-        self._memory_reader_manager.update_process_target(target_payload)
-        with self._state_lock:
-            self._state.memory_reader_target = json_copy(target_payload)
-            self._state_dirty = True
-            self._cached_snapshot = None
-        background_poll_started = self._start_background_bridge_poll()
-        return Ok(
-            {
-                "process_target": json_copy(target_payload),
-                "cleared": bool(clear),
-                "summary": summary,
-                "background_poll_started": background_poll_started,
-            }
-        )
-
-    @plugin_entry(
-        id="galgame_list_ocr_windows",
-        name=tr("entries.galgame_list_ocr_windows.name", default='列出 OCR 候选窗口'),
-        description=tr("entries.galgame_list_ocr_windows.description", default='返回当前 OCR Reader 的可选窗口，可选包含只读排除列表。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "include_excluded": {"type": "boolean", "default": False},
-                "force": {"type": "boolean", "default": False},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_list_ocr_windows(
-        self,
-        include_excluded: bool = False,
-        force: bool = False,
-        **_,
-    ):
-        if self._ocr_reader_manager is None:
-            return Err(SdkError("ocr_reader manager is not initialized"))
-        try:
-            payload = await asyncio.to_thread(
-                self._ocr_reader_manager.list_windows_snapshot,
-                include_excluded=bool(include_excluded),
-                force=bool(force),
-            )
-        except Exception as exc:
-            return Err(SdkError(f"list OCR windows failed: {exc}"))
-        payload["summary"] = (
-            f"eligible={int(payload.get('candidate_count') or 0)} "
-            f"excluded={int(payload.get('excluded_candidate_count') or 0)} "
-            f"mode={payload.get('target_selection_mode') or 'auto'}"
-        )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_set_ocr_window_target",
-        name=tr("entries.galgame_set_ocr_window_target.name", default='设置 OCR 目标窗口'),
-        description=tr("entries.galgame_set_ocr_window_target.description", default='锁定或清除 OCR Reader 的手动目标窗口。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "window_key": {"type": "string", "default": ""},
-                "clear": {"type": "boolean", "default": False},
-            },
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_ocr_window_target(
-        self,
-        window_key: str = "",
-        clear: bool = False,
-        **_,
-    ):
-        if self._ocr_reader_manager is None:
-            return Err(SdkError("ocr_reader manager is not initialized"))
-
-        if clear:
-            target_payload = {
-                "mode": "auto",
-                "window_key": "",
-                "process_name": "",
-                "normalized_title": "",
-                "pid": 0,
-                "last_known_hwnd": 0,
-                "selected_at": "",
-            }
-            summary = "OCR window target cleared; waiting for manual lock"
-        else:
-            try:
-                target_payload = await asyncio.to_thread(
-                    self._ocr_reader_manager.resolve_manual_window_target,
-                    window_key,
-                )
-            except ValueError as exc:
-                return Err(SdkError(str(exc)))
-            except Exception as exc:
-                return Err(SdkError(f"resolve OCR window target failed: {exc}"))
-            summary = (
-                f"OCR window target locked to {target_payload.get('process_name') or '(unknown)'}"
-            )
-
-        try:
-            self._persist.persist_ocr_window_target(target_payload)
-        except Exception as exc:
-            return Err(SdkError(f"persist OCR window target failed: {exc}"))
-
-        with self._state_lock:
-            self._state.ocr_window_target = json_copy(target_payload)
-            self._state_dirty = True
-            self._cached_snapshot = None
-        self._ocr_reader_manager.update_window_target(target_payload)
-        background_poll_started = self._start_background_bridge_poll()
-        return Ok(
-            {
-                "window_target": json_copy(target_payload),
-                "cleared": bool(clear),
-                "summary": summary,
-                "background_poll_started": background_poll_started,
-            }
-        )
-
-    @plugin_entry(
-        id="galgame_open_ui",
-        name=tr("entries.galgame_open_ui.name", default='打开 galgame UI'),
-        description=tr("entries.galgame_open_ui.description", default='返回 galgame_plugin 静态 UI 的访问路径。'),
-        input_schema={"type": "object", "properties": {}},
-        llm_result_fields=["message"],
-    )
-    async def galgame_open_ui(self, **_):
-        payload = build_open_ui_payload(
-            plugin_id=self.plugin_id,
-            available=self.get_static_ui_config() is not None,
-        )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_explain_line",
-        name=tr("entries.galgame_explain_line.name", default='解释当前或指定台词'),
-        description=tr("entries.galgame_explain_line.description", default='对当前快照或指定 line_id 对应的台词进行解释。'),
-        input_schema={
-            "type": "object",
-            "properties": {"line_id": {"type": "string", "default": ""}},
-        },
-        timeout=45.0,
-        llm_result_fields=["explanation", "diagnostic"],
-    )
-    async def galgame_explain_line(self, line_id: str = "", **_):
-        if self._llm_gateway is None:
-            return Err(SdkError("galgame_plugin llm_gateway is not initialized"))
-        local = self._snapshot_state(include_private_context=True)
-        try:
-            context = build_explain_context(local, line_id=line_id.strip(), config=self._cfg)
-        except ValueError as exc:
-            context = {
-                "line_id": "",
-                "speaker": "",
-                "text": "",
-                "scene_id": "",
-                "route_id": "",
-                "evidence": [],
-            }
-            return Ok(
-                build_explain_degraded_result(
-                    context,
-                    diagnostic=str(exc) or build_ocr_context_diagnostic(local),
-                )
-            )
-        payload = apply_input_degraded_result(
-            await self._llm_gateway.explain_line(context),
-            context=context,
-        )
-        payload["line_id"] = str(context.get("line_id") or "")
-        payload["speaker"] = str(context.get("speaker") or "")
-        payload["text"] = str(context.get("text") or "")
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_summarize_scene",
-        name=tr("entries.galgame_summarize_scene.name", default='总结当前场景'),
-        description=tr("entries.galgame_summarize_scene.description", default='总结当前场景或指定 scene_id 的最近剧情进展。'),
-        input_schema={
-            "type": "object",
-            "properties": {"scene_id": {"type": "string", "default": ""}},
-        },
-        timeout=45.0,
-        llm_result_fields=["summary", "diagnostic"],
-    )
-    async def galgame_summarize_scene(self, scene_id: str = "", **_):
-        if self._llm_gateway is None:
-            return Err(SdkError("galgame_plugin llm_gateway is not initialized"))
-        local = self._snapshot_state(include_private_context=True)
-        context = build_summarize_context(local, scene_id=scene_id.strip(), config=self._cfg)
-        snapshot = context.get("current_snapshot") if isinstance(context.get("current_snapshot"), dict) else {}
-        if not list(context.get("recent_lines") or []) and not str(snapshot.get("text") or ""):
-            return Ok(
-                build_summarize_degraded_result(
-                    context,
-                    diagnostic=build_ocr_context_diagnostic(local),
-                )
-            )
-        payload = apply_input_degraded_result(
-            await self._llm_gateway.summarize_scene(context),
-            context=context,
-        )
-        payload["scene_id"] = str(context.get("scene_id") or "")
-        try:
-            await asyncio.to_thread(self._persist_context_snapshot_from_summary, context, payload)
-        except Exception as exc:
-            self.logger.warning(
-                "persist context snapshot from scene summary failed: {}",
-                exc,
-            )
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_suggest_choice",
-        name=tr("entries.galgame_suggest_choice.name", default='建议当前选项'),
-        description=tr("entries.galgame_suggest_choice.description", default='对当前可见选项给出推荐顺位与理由。'),
-        input_schema={"type": "object", "properties": {}},
-        timeout=45.0,
-        llm_result_fields=["choices", "diagnostic"],
-    )
-    async def galgame_suggest_choice(self, **_):
-        if self._llm_gateway is None:
-            return Err(SdkError("galgame_plugin llm_gateway is not initialized"))
-        local = self._snapshot_state(include_private_context=True)
-        context = build_suggest_context(local, config=self._cfg)
-        if not context["visible_choices"]:
-            return Ok(
-                apply_input_degraded_result(
-                    build_suggest_degraded_result(
-                        context,
-                        diagnostic="gateway_unavailable: no visible choices",
-                    ),
-                    context=context,
-                )
-            )
-        payload = apply_input_degraded_result(
-            await self._llm_gateway.suggest_choice(context),
-            context=context,
-        )
-        payload["scene_id"] = str(context.get("scene_id") or "")
-        return Ok(payload)
-
-    @plugin_entry(
-        id="galgame_agent_command",
-        name=tr("entries.galgame_agent_command.name", default='向 Game LLM Agent 发送指令'),
-        description=tr("entries.galgame_agent_command.description", default='查询 Agent 状态、上下文、发送消息或控制待机。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "action": {
-                    "type": "string",
-                    "enum": [
-                        "query_status",
-                        "query_context",
-                        "send_message",
-                        "set_standby",
-                        "list_messages",
-                        "ack_message",
-                    ],
-                },
-                "message": {"type": "string", "default": ""},
-                "context_query": {"type": "string", "default": ""},
-                "message_id": {"type": "string", "default": ""},
-                "reply_to_message_id": {"type": "string", "default": ""},
-                "sender_role": {"type": "string", "default": ""},
-                "consultation_reply": {"type": "boolean", "default": False},
-                "direction": {"type": "string", "default": ""},
-                "limit": {"type": "integer", "default": 50},
-                "standby": {"type": "boolean"},
-            },
-            "required": ["action"],
-        },
-        timeout=45.0,
-        llm_result_fields=["result", "status"],
-    )
-    async def galgame_agent_command(
-        self,
-        action: str,
-        message: str = "",
-        context_query: str = "",
-        message_id: str = "",
-        reply_to_message_id: str = "",
-        sender_role: str = "",
-        consultation_reply: bool = False,
-        direction: str = "",
-        limit: int = 50,
-        standby: bool | None = None,
-        **_,
-    ):
-        if self._game_agent is None:
-            return Err(SdkError("galgame_plugin game agent is not initialized"))
-        local = self._snapshot_state(include_private_context=True)
-        if action == "query_status":
-            return Ok(await self._game_agent.query_status(local))
-        if action == "query_context":
-            if not context_query.strip():
-                return Err(SdkError("context_query is required for query_context"))
-            return Ok(
-                await self._game_agent.query_context(
-                    local,
-                    context_query=context_query.strip(),
-                )
-            )
-        if action == "send_message":
-            if not message.strip():
-                return Err(SdkError("message is required for send_message"))
-            return Ok(
-                await self._game_agent.send_message(
-                    local,
-                    message=message.strip(),
-                    reply_to_message_id=reply_to_message_id.strip(),
-                    sender_role=sender_role.strip(),
-                    consultation_reply=bool(consultation_reply),
-                )
-            )
-        if action == "set_standby":
-            if standby is None:
-                return Err(SdkError("standby is required for set_standby"))
-            return Ok(await self._game_agent.set_standby(local, standby=bool(standby)))
-        if action == "list_messages":
-            return Ok(
-                await self._game_agent.list_messages(
-                    local,
-                    direction=direction,
-                    limit=int(limit or 50),
-                )
-            )
-        if action == "ack_message":
-            if not message_id.strip():
-                return Err(SdkError("message_id is required for ack_message"))
-            return Ok(
-                await self._game_agent.ack_message(
-                    local,
-                    message_id=message_id.strip(),
-                )
-            )
-        return Err(SdkError(f"unsupported agent action: {action!r}"))
-
-    @plugin_entry(
-        id="galgame_continue_auto_advance",
-        name=tr("entries.galgame_continue_auto_advance.name", default='继续自动推进 galgame 剧情'),
-        description=tr("entries.galgame_continue_auto_advance.description", default='切换到自动推进模式，并向 Game LLM Agent 发送继续推进消息。'),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "message": {"type": "string", "default": "继续推进剧情"},
-            },
-        },
-        timeout=45.0,
-        llm_result_fields=[
-            "result",
-            "status",
-            "mode",
-            "mode_result",
-            "agent_result",
-            "diagnostic",
-        ],
-    )
-    async def galgame_continue_auto_advance(
-        self,
-        message: str = "继续推进剧情",
-        **_,
-    ):
-        normalized_message = str(message or "").strip() or "继续推进剧情"
-        mode_res = await self.galgame_set_mode(
-            mode="choice_advisor",
-            push_notifications=True,
-        )
-        if isinstance(mode_res, Err):
-            return mode_res
-        mode_payload = json_copy(mode_res.value or {})
-
-        agent_res = await self.galgame_agent_command(
-            action="send_message",
-            message=normalized_message,
-        )
-        if isinstance(agent_res, Err):
-            return Err(
-                SdkError(
-                    f"continue auto advance send_message failed: {agent_res.error}",
-                    details={
-                        "mode_result": mode_payload,
-                        "message": normalized_message,
-                    },
-                    mode_result=mode_payload,
-                )
-            )
-        agent_payload = json_copy(agent_res.value or {})
-        status = (
-            json_copy(agent_payload.get("status"))
-            if isinstance(agent_payload, dict)
-            else {}
-        )
-        result_text = (
-            str(agent_payload.get("result") or "")
-            if isinstance(agent_payload, dict)
-            else ""
-        )
-        diagnostic = (
-            str(agent_payload.get("diagnostic") or "")
-            if isinstance(agent_payload, dict)
-            else ""
-        )
-        return Ok(
-            {
-                "action": "continue_auto_advance",
-                "message": normalized_message,
-                "mode": "choice_advisor",
-                "mode_result": {
-                    "success": True,
-                    "mode": "choice_advisor",
-                    "push_notifications": True,
-                    "result": mode_payload,
-                },
-                "agent_result": agent_payload,
-                "status": status,
-                "result": result_text,
-                "degraded": False,
-                "diagnostic": diagnostic,
-            }
-        )
 
 
     # ------------------------------------------------------------------
@@ -6509,455 +4755,17 @@ class GalgamePlugin(NekoPluginBase):
                 int(push_seq or 0),
             )
 
-    @plugin_entry(
-        id="galgame_get_scene_context",
-        name=tr(
-            "entries.galgame_get_scene_context.name",
-            default="查询场景上下文",
-        ),
-        description=tr(
-            "entries.galgame_get_scene_context.description",
-            default="获取当前或指定场景的完整上下文摘要 + 关键台词。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "scene_id": {"type": "string", "default": ""},
-            },
-        },
-        timeout=5.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_scene_context(
-        self,
-        scene_id: str = "",
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        throttled = self._check_query_rate_limit("galgame_get_scene_context")
-        if throttled is not None:
-            return Ok(throttled)
-        scenes = self._layer1_scene_summaries()
-        target = (scene_id or "").strip()
-        entry: dict[str, Any] | None = None
-        if target:
-            for item in reversed(scenes):
-                if str(item.get("scene_id") or "") == target:
-                    entry = item
-                    break
-        elif scenes:
-            entry = scenes[-1]
-        if entry is None:
-            return Ok(
-                {
-                    "scene_id": target,
-                    "found": False,
-                    "summary": "no scene context available",
-                }
-            )
-        return Ok(
-            {
-                "scene_id": str(entry.get("scene_id") or ""),
-                "route_id": str(entry.get("route_id") or ""),
-                "found": True,
-                "summary_text": str(entry.get("summary") or ""),
-                "key_lines": list(entry.get("key_lines") or []),
-                "ts": entry.get("ts") or "",
-                "summary": (
-                    f"scene={entry.get('scene_id') or '-'} "
-                    f"lines={len(entry.get('key_lines') or [])}"
-                ),
-            }
-        )
 
-    @plugin_entry(
-        id="galgame_get_story_so_far",
-        name=tr(
-            "entries.galgame_get_story_so_far.name",
-            default="查询全局故事线",
-        ),
-        description=tr(
-            "entries.galgame_get_story_so_far.description",
-            default="获取从游戏开始到现在的完整故事线摘要（约 200 tokens）。",
-        ),
-        input_schema={"type": "object", "properties": {}},
-        timeout=5.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_story_so_far(self, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        throttled = self._check_query_rate_limit("galgame_get_story_so_far")
-        if throttled is not None:
-            return Ok(throttled)
-        self._refresh_story_so_far_from_scene_summaries()
-        story = (self._story_so_far or "").strip()
-        if not story:
-            return Ok(
-                {
-                    "story_so_far": "故事刚开始。",
-                    "available": False,
-                    "summary": "no story summary yet",
-                }
-            )
-        return Ok(
-            {
-                "story_so_far": story,
-                "available": True,
-                "last_updated_seq": int(self._story_last_updated_seq or 0),
-                "summary": f"story summary ({len(story)} chars)",
-            }
-        )
 
-    @plugin_entry(
-        id="galgame_get_recent_lines",
-        name=tr(
-            "entries.galgame_get_recent_lines.name",
-            default="查询最近台词",
-        ),
-        description=tr(
-            "entries.galgame_get_recent_lines.description",
-            default="获取最近 N 句原始台词内容（上限 20）。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "n": {"type": "integer", "default": 10, "minimum": 1, "maximum": 20},
-            },
-        },
-        timeout=5.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_recent_lines(self, n: int = 10, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        throttled = self._check_query_rate_limit("galgame_get_recent_lines")
-        if throttled is not None:
-            return Ok(throttled)
-        try:
-            requested = int(n)
-        except (TypeError, ValueError):
-            requested = 10
-        requested = max(1, min(20, requested))
-        with self._state_lock:
-            history = [dict(line) for line in (self._state.history_lines or [])]
-        recent = history[-requested:]
-        return Ok(
-            {
-                "lines": recent,
-                "count": len(recent),
-                "summary": f"{len(recent)} line(s) returned",
-            }
-        )
 
-    @plugin_entry(
-        id="galgame_get_push_history",
-        name=tr(
-            "entries.galgame_get_push_history.name",
-            default="查询推送历史",
-        ),
-        description=tr(
-            "entries.galgame_get_push_history.description",
-            default="返回最近 N 次推送的元数据（push_seq/mode/token/场景/角色），不含台词文本。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "limit": {"type": "integer", "default": 10, "minimum": 1, "maximum": 20},
-            },
-        },
-        timeout=5.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_push_history(self, limit: int = 10, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        throttled = self._check_query_rate_limit("galgame_get_push_history")
-        if throttled is not None:
-            return Ok(throttled)
-        try:
-            requested = int(limit)
-        except (TypeError, ValueError):
-            requested = 10
-        requested = max(1, min(20, requested))
-        history = list(self._push_history)[-requested:]
-        return Ok(
-            {
-                "entries": history,
-                "count": len(history),
-                "summary": f"{len(history)} push record(s) returned",
-            }
-        )
 
     # ------------------------------------------------------------------
     # Character profile entries (galgame-host-play-mode plan, step 6)
     # ------------------------------------------------------------------
 
-    @plugin_entry(
-        id="galgame_set_character_mode",
-        name=tr(
-            "entries.galgame_set_character_mode.name",
-            default="设置角色档案模式",
-        ),
-        description=tr(
-            "entries.galgame_set_character_mode.description",
-            default="切换角色档案模式：off 或 fixed（锁定某角色作为猫娘视角）。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "mode": {"type": "string", "enum": ["off", "fixed"]},
-                "character_name": {"type": "string", "default": ""},
-            },
-            "required": ["mode"],
-        },
-        llm_result_fields=["summary"],
-    )
-    async def galgame_set_character_mode(
-        self,
-        mode: str,
-        character_name: str = "",
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        mode_normalized = (mode or "").strip().lower()
-        if mode_normalized not in {"off", "fixed"}:
-            return Err(SdkError(f"invalid character mode: {mode!r}"))
 
-        with self._state_lock:
-            current_profiles = dict(self._state.character_profiles or {})
 
-        if mode_normalized == "fixed":
-            name = (character_name or "").strip()
-            if not name:
-                return Err(SdkError("fixed mode requires character_name"))
-            load = self._load_character_profiles_for_current_context()
-            loaded_profiles = dict(load.get("profiles") or {})
-            if loaded_profiles:
-                current_profiles = loaded_profiles
-            if not current_profiles:
-                current_profiles = dict(load.get("profiles") or {})
-                errors = load.get("errors") or []
-                resolved_game_id = str(load.get("resolved_game_id") or "")
-                return Err(
-                    SdkError(
-                        "no character profiles available"
-                        + (f" for {resolved_game_id}" if resolved_game_id else ""),
-                        details={"errors": list(errors)},
-                    )
-                )
-            if name not in current_profiles:
-                return Err(SdkError(f"character {name!r} not found in profiles"))
 
-        with self._state_lock:
-            self._state.character_mode = mode_normalized
-            self._state.character_fixed_name = (
-                (character_name or "").strip()
-                if mode_normalized == "fixed"
-                else ""
-            )
-            self._state.character_mode_stale = False
-            self._state_dirty = True
-            self._cached_snapshot = None
-            fixed_name = str(self._state.character_fixed_name or "")
-
-        try:
-            self._persist.persist_config_override(
-                STORE_CHARACTER_MODE, mode_normalized
-            )
-            self._persist.persist_config_override(
-                STORE_CHARACTER_FIXED_NAME, fixed_name
-            )
-        except Exception:  # noqa: BLE001
-            self.logger.warning(
-                "failed to persist character mode switch", exc_info=True
-            )
-
-        return Ok(
-            {
-                "mode": mode_normalized,
-                "character_name": fixed_name,
-                "summary": (
-                    f"character_mode={mode_normalized} character={fixed_name or '-'}"
-                ),
-            }
-        )
-
-    @plugin_entry(
-        id="galgame_get_character_list",
-        name=tr(
-            "entries.galgame_get_character_list.name",
-            default="查询当前游戏角色列表",
-        ),
-        description=tr(
-            "entries.galgame_get_character_list.description",
-            default="返回当前绑定游戏的可用角色名称和一句话身份描述。",
-        ),
-        input_schema={"type": "object", "properties": {}},
-        timeout=10.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_character_list(self, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        with self._state_lock:
-            bound_game_id = str(self._state.bound_game_id or "")
-            profile_game_id = str(self._state.character_profile_game_id or "")
-            match_reason = str(self._state.character_profile_match_reason or "")
-            profiles = dict(self._state.character_profiles or {})
-        load = self._load_character_profiles_for_current_context(force=True)
-        loaded_profiles = dict(load.get("profiles") or {})
-        if loaded_profiles or not profiles:
-            profiles = loaded_profiles
-            profile_game_id = str(load.get("resolved_game_id") or "")
-            match_reason = str(load.get("match_reason") or "")
-        items = [
-            {
-                "name": name,
-                "identity": str((profile or {}).get("identity") or ""),
-            }
-            for name, profile in profiles.items()
-        ]
-        return Ok(
-            {
-                "game_id": profile_game_id or bound_game_id,
-                "profile_game_id": profile_game_id,
-                "match_reason": match_reason,
-                "characters": items,
-                "summary": f"{len(items)} character(s) available",
-            }
-        )
-
-    @plugin_entry(
-        id="galgame_get_character_profile",
-        name=tr(
-            "entries.galgame_get_character_profile.name",
-            default="查询角色完整档案",
-        ),
-        description=tr(
-            "entries.galgame_get_character_profile.description",
-            default="返回指定角色的预置档案 + 运行时状态。off 模式下拒绝。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "name": {"type": "string"},
-            },
-            "required": ["name"],
-        },
-        timeout=10.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_get_character_profile(self, name: str, **_):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        with self._state_lock:
-            mode = str(self._state.character_mode or "off")
-            profiles = dict(self._state.character_profiles or {})
-            runtime = dict(self._state.character_runtime_state or {})
-        if mode == "off":
-            return Ok(
-                {
-                    "disabled": True,
-                    "reason": (
-                        "角色档案功能未开启，请先调用 galgame_set_character_mode"
-                        " 切换为 fixed"
-                    ),
-                    "summary": "character_mode=off",
-                }
-            )
-        target = (name or "").strip()
-        if not target:
-            return Err(SdkError("character name required"))
-        if not profiles:
-            load = self._load_character_profiles_for_current_context()
-            profiles = dict(load.get("profiles") or {})
-            with self._state_lock:
-                runtime = dict(self._state.character_runtime_state or {})
-        profile = profiles.get(target)
-        if not profile:
-            return Err(SdkError(f"character {target!r} not found"))
-        return Ok(
-            {
-                "name": target,
-                "profile": profile,
-                "runtime_state": runtime.get(target, {}),
-                "summary": f"profile for {target}",
-            }
-        )
-
-    @plugin_entry(
-        id="galgame_import_character_data",
-        name=tr(
-            "entries.galgame_import_character_data.name",
-            default="导入角色档案 JSON",
-        ),
-        description=tr(
-            "entries.galgame_import_character_data.description",
-            default="从 JSON 文件导入用户自定义角色档案。",
-        ),
-        input_schema={
-            "type": "object",
-            "properties": {
-                "file_path": {"type": "string"},
-                "game_id": {"type": "string", "default": ""},
-            },
-            "required": ["file_path"],
-        },
-        timeout=15.0,
-        llm_result_fields=["summary"],
-    )
-    async def galgame_import_character_data(
-        self,
-        file_path: str,
-        game_id: str = "",
-        **_,
-    ):
-        if self._cfg is None:
-            return Err(SdkError(self._not_configured_message()))
-        target_game = (game_id or "").strip()
-        if not target_game:
-            with self._state_lock:
-                target_game = str(self._state.bound_game_id or "")
-        if not target_game:
-            return Err(
-                SdkError("game_id required (no game currently bound)")
-            )
-        source = (file_path or "").strip()
-        if not source:
-            return Err(SdkError("file_path required"))
-        manager = self._get_character_profile_manager()
-        result = manager.import_user_profiles(target_game, source)
-        if not result.ok:
-            return Err(
-                SdkError(
-                    f"import failed: {'; '.join(result.errors)}",
-                    details={
-                        "target_path": result.target_path,
-                        "errors": list(result.errors),
-                        "warnings": list(result.warnings),
-                    },
-                )
-            )
-        # Reload merged profiles only when the import targets the active binding.
-        with self._state_lock:
-            active_game = str(self._state.bound_game_id or self._state.active_game_id or "")
-        if target_game == active_game:
-            self._activate_character_profiles(target_game)
-        return Ok(
-            {
-                "ok": True,
-                "game_id": target_game,
-                "target_path": result.target_path,
-                "profile_count": result.profile_count,
-                "warnings": list(result.warnings),
-                "summary": (
-                    f"imported {result.profile_count} profile(s) for {target_game}"
-                ),
-            }
-        )
 
 
 GalgameBridgePlugin = GalgamePlugin
